Reuse GQL fragment AWSVPC
