+++ reconcile/gql_definitions/fragments/aws_vpc.gql
@@ -6,8 +6,15 @@ fragment AWSVPC on AWSVPC_v1 {
   account {
     name
     uid
+    terraformUsername
+    automationToken {
+      ... VaultSecret
+    }
   }
   region
   vpc_id
   cidr_block
+  subnets {
+    id
+  }
 }
